fix: update direction count validation to assert an expected count
This way, if the number of directions for a route changes in the future,
we'll be alerted to it as a validation failure and can make the decision to
change the expectation.

For example, route 214 has only one direction in the Spring 2021 rating. If
it goes back to two directions in Summer or Fall, this validation will fail
and we can adjust it.