fix: raise a better error when Universal dir isn't found
We look for the current ANN2DEST file in the Universal announcements on
the TransitMaster DB server. Previously, if we didn't find it, we'd
raise a cryptic `IndexError`. Now, we raise a friendlier `RuntimeError`:

```
RuntimeError: unable to find Universal announcements in \\hstmcldb\e$\FTP_ROOT\Operational Data\Announcements\Current_Release
Found: ['071522_BRND']
```